+++ Source/ACE/Entity/GameData.cs
@@ -41,7 +41,7 @@ namespace ACE.Entity
         public ushort Script;
         public float Workmanship;
         public ushort Burden;
-        public ushort Spell;
+        public Spell Spell;
 
         // Housing links to another packet, that needs sent.. The HouseRestrictions ACL Control list that contains all the housing data
         public uint HouseOwner;